Add search related endpoints
