+++ src/main/java/com/example/mentormenteewebsite/MentorMenteeWebsiteApplication.java
@@ -2,8 +2,9 @@ package com.example.mentormenteewebsite;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {JacksonAutoConfiguration.class})
 public class MentorMenteeWebsiteApplication {
 
 	public static void main(String[] args) {
