+++ src/main/java/com/example/mentormenteewebsite/database/profileMicroservice/MentorRepository.java
@@ -1,6 +1,6 @@
-package com.example.mentormenteewebsite.database;
+package com.example.mentormenteewebsite.database.profileMicroservice;
 
-import com.example.mentormenteewebsite.models.Mentor;
+import com.example.mentormenteewebsite.models.profileMicroservice.Mentor;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
