+++ src/main/java/com/example/mentormenteewebsite/database/profileMicroservice/UserRepositiory.java
@@ -1,10 +1,9 @@
-package com.example.mentormenteewebsite.database;
+package com.example.mentormenteewebsite.database.profileMicroservice;
 
-import com.example.mentormenteewebsite.models.User;
+import com.example.mentormenteewebsite.models.profileMicroservice.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface UserRepositiory extends MongoRepository<User, String> {
     List<User> findAll();
