--- src/main/java/com/example/mentormenteewebsite/models/MentorType.java
@@ -1,5 +1,0 @@
-package com.example.mentormenteewebsite.models;
-
-public enum MentorType {
-    Software
-}
