--- src/main/java/com/example/mentormenteewebsite/models/UserType.java
@@ -1,6 +1,0 @@
-package com.example.mentormenteewebsite.models;
-
-public enum UserType {
-    Mentor,
-    Mentee
-}
