+++ src/main/java/com/example/mentormenteewebsite/models/profileMicroservice/Experties.java
@@ -1,10 +1,13 @@
-package com.example.mentormenteewebsite.models;
+package com.example.mentormenteewebsite.models.profileMicroservice;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.Arrays;
 import java.util.List;
 
+@Document(collection = "experties")
 public class Experties {
     private List<String> experties;
 
@@ -15,18 +18,18 @@ public class Experties {
         this.experties = experties;
     }
 
-    @Override
-    public String toString() {
-        return "Experties{" +
-                "experties=" + experties +
-                '}';
-    }
-
     public List<String> getExperties() {
         return experties;
     }
 
     public void setExperties(List<String> experties) {
         this.experties = experties;
+    }
+
+    @Override
+    public String toString() {
+        return "Experties{" +
+                "experties=" + experties +
+                '}';
     }
 }
