+++ src/main/java/com/example/mentormenteewebsite/models/profileMicroservice/MentorType.java
@@ -1,0 +1,5 @@
+package com.example.mentormenteewebsite.models.profileMicroservice;
+
+public enum MentorType {
+    Software
+}
