+++ src/main/java/com/example/mentormenteewebsite/models/profileMicroservice/User.java
@@ -1,6 +1,5 @@
-package com.example.mentormenteewebsite.models;
+package com.example.mentormenteewebsite.models.profileMicroservice;
 
-import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "User")
