+++ src/main/java/com/example/mentormenteewebsite/models/profileMicroservice/UserType.java
@@ -1,0 +1,6 @@
+package com.example.mentormenteewebsite.models.profileMicroservice;
+
+public enum UserType {
+    Mentor,
+    Mentee
+}
