+++ src/main/resources/application.properties
@@ -2,3 +2,4 @@
 spring.data.mongodb.database=mentorWebsite
 spring.data.mongodb.port=27017
 spring.data.mongodb.host=localhost
+spring.http.converters.preferred-json-mapper=gson
