+++ src/test/java/controller/ProfileManagementTest.java
@@ -1,6 +1,6 @@
 package controller;
 
-import com.example.mentormenteewebsite.controller.ProfileManagement;
+import com.example.mentormenteewebsite.controller.profileMicroservice.ProfileManagement;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
